requirements.txt
Django==5.2.4
django-bootstrap5==25.2
django-mptt==0.17.0
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.4
